make everything compile -- some tests fail do to changing geometry and molecule
